backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
django-cors-headers>=4.3

djangorestframework
djangorestframework-simplejwt
pdfplumber
requests>=2.31
